+++ src/Integration.Vsix.UnitTests/CFamily/CFamilyHelperTests.cs
@@ -31,6 +31,7 @@ using SonarLint.VisualStudio.Core.Analysis;
 using SonarLint.VisualStudio.Core.CFamily;
 using SonarLint.VisualStudio.Integration.UnitTests;
 using SonarLint.VisualStudio.Integration.UnitTests.CFamily;
+using static SonarLint.VisualStudio.Integration.Vsix.CFamily.UnitTests.CFamilyTestUtility;
 
 namespace SonarLint.VisualStudio.Integration.Vsix.CFamily.UnitTests
 {
@@ -189,60 +190,6 @@ namespace SonarLint.VisualStudio.Integration.Vsix.CFamily.UnitTests
         }
 
         [TestMethod]
-        public void CreateRequest_UnsupportedItemType()
-        {
-            // Arrange
-            var loggerMock = new Mock<ILogger>();
-            ProjectItemConfig projectItemConfig = new ProjectItemConfig();
-            projectItemConfig.itemType = "None";
-            var projectItemMock = CreateProjectItemWithProject("c:\\foo\\xxx.vcxproj", projectItemConfig);
-            // Act
-            var request = CFamilyHelper.CreateRequest(loggerMock.Object, projectItemMock.Object, "c:\\dummy\\file.cpp",
-                null, null);
-
-            // Assert
-            AssertMessageLogged(loggerMock,
-                "File's \"Item type\" is not supported. File: 'c:\\dummy\\file.cpp'");
-            request.Should().BeNull();
-        }
-
-        [TestMethod]
-        public void CreateRequest_UnsupportedConfigurationType()
-        {
-            // Arrange
-            var loggerMock = new Mock<ILogger>();
-            ProjectItemConfig projectItemConfig = new ProjectItemConfig();
-            projectItemConfig.configurationType = ConfigurationTypes.typeUnknown;
-            var projectItemMock = CreateProjectItemWithProject("c:\\foo\\xxx.vcxproj", projectItemConfig);
-            // Act
-            var request = CFamilyHelper.CreateRequest(loggerMock.Object, projectItemMock.Object, "c:\\dummy\\file.cpp",
-                null, null);
-
-            // Assert
-            AssertMessageLogged(loggerMock,
-                "Project's \"Configuration type\" is not supported.");
-            request.Should().BeNull();
-        }
-
-        [TestMethod]
-        public void CreateRequest_UnsupportedCustomBuild()
-        {
-            // Arrange
-            var loggerMock = new Mock<ILogger>();
-            ProjectItemConfig projectItemConfig = new ProjectItemConfig();
-            projectItemConfig.isVCCLCompilerTool = false;
-            var projectItemMock = CreateProjectItemWithProject("c:\\foo\\xxx.vcxproj", projectItemConfig);
-            // Act
-            var request = CFamilyHelper.CreateRequest(loggerMock.Object, projectItemMock.Object, "c:\\dummy\\file.cpp",
-                null, null);
-
-            // Assert
-            AssertMessageLogged(loggerMock,
-                "Custom built files are not supported. File: 'c:\\dummy\\file.cpp'");
-            request.Should().BeNull();
-        }
-
-        [TestMethod]
         public void IsFileInSolution_NullItem_ReturnsFalse()
         {
             // Arrange and Act
@@ -372,59 +319,6 @@ namespace SonarLint.VisualStudio.Integration.Vsix.CFamily.UnitTests
             return config;
         }
 
-        class ProjectItemConfig
-        {
-            public string platformName { get; set; } = "Win32";
-            public IDictionary<string, string> projectConfigProperties { get; set; } = new Dictionary<string, string>
-            {
-                ["PlatformToolset"] = "v140_xp"
-            };
-            public IDictionary<string, string> fileConfigProperties { get; set; } = new Dictionary<string, string>
-            {
-                ["PrecompiledHeader"] = "NotUsing",
-                ["CompileAs"] = "CompileAsCpp",
-                ["CompileAsManaged"] = "false",
-                ["EnableEnhancedInstructionSet"] = "",
-                ["RuntimeLibrary"] = "",
-                ["LanguageStandard"] = "",
-                ["ExceptionHandling"] = "Sync",
-                ["BasicRuntimeChecks"] = "UninitializedLocalUsageCheck",
-            };
-            public bool isVCCLCompilerTool { get; set; } = true;
-            public string itemType { get; set; } = "ClCompile";
-            public ConfigurationTypes configurationType { get; set; } = ConfigurationTypes.typeApplication;
-        }
-
-        private static readonly ProjectItemConfig defaultSetting = new ProjectItemConfig();
-
-        private Mock<ProjectItem> CreateProjectItemWithProject(string projectName, ProjectItemConfig projectItemConfig = null)
-        {
-            projectItemConfig = projectItemConfig ?? defaultSetting;
-
-            var vcProjectMock = new Mock<VCProject>();
-            var vcConfig = CreateVCConfigurationWithProperties(projectItemConfig);
-            vcProjectMock.SetupGet(x => x.ActiveConfiguration).Returns(vcConfig);
-
-            var projectMock = new ProjectMock(projectName) {Project = vcProjectMock.Object};
-
-            var vcFileMock = new Mock<VCFile>();
-            vcFileMock.SetupGet(x => x.ItemType).Returns(projectItemConfig.itemType);
-            var vcFileConfig = CreateVCFileConfigurationWithToolProperties(projectItemConfig);
-            vcFileMock.Setup(x => x.GetFileConfigurationForProjectConfiguration(vcConfig)).Returns(vcFileConfig);
-            var projectItemMock = new Mock<ProjectItem>();
-            projectItemMock.Setup(i => i.ContainingProject).Returns(projectMock);
-            projectItemMock.Setup(i => i.Object).Returns(vcFileMock.Object);
-
-            // Set the project item to have a valid DTE configuration
-            // - used to check whether the project item is in a solution or not
-            var dteConfigManagerMock = new Mock<ConfigurationManager>();
-            var dteConfigMock = new Mock<Configuration>();
-            dteConfigManagerMock.Setup(x => x.ActiveConfiguration).Returns(dteConfigMock.Object);
-            projectItemMock.Setup(i => i.ConfigurationManager).Returns(dteConfigManagerMock.Object);
-
-            return projectItemMock;
-        }
-
         private Request GetSuccessfulRequest(IAnalyzerOptions analyzerOptions)
         {
             var loggerMock = new Mock<ILogger>();
@@ -440,60 +334,6 @@ namespace SonarLint.VisualStudio.Integration.Vsix.CFamily.UnitTests
                 rulesConfigProviderMock.Object, analyzerOptions);
 
             return request;
-        }
-
-        private static VCConfiguration CreateVCConfigurationWithProperties(ProjectItemConfig projectItemConfig)
-        {
-            var vcPlatformMock = new Mock<VCPlatform>();
-            vcPlatformMock.SetupGet(x => x.Name).Returns(projectItemConfig.platformName);
-
-            var vcConfigMock = new Mock<VCConfiguration>();
-            vcConfigMock.SetupGet(x => x.Platform).Returns(vcPlatformMock.Object);
-            vcConfigMock.SetupGet(x => x.ConfigurationType).Returns(projectItemConfig.configurationType);
-
-            vcConfigMock.Setup(x => x.GetEvaluatedPropertyValue(It.IsAny<string>()))
-                .Returns<string>(s =>
-                {
-                    string propertyValue = null;
-                    projectItemConfig.projectConfigProperties?.TryGetValue(s, out propertyValue);
-                    return propertyValue ?? string.Empty;
-                });
-
-            return vcConfigMock.Object;
-        }
-
-        private static VCFileConfiguration CreateVCFileConfigurationWithToolProperties(ProjectItemConfig projectItemConfig)
-        {
-            var toolPropertiesMock = new Mock<IVCRulePropertyStorage>();
-            if (projectItemConfig.isVCCLCompilerTool)
-            {
-                toolPropertiesMock.As<VCCLCompilerTool>();
-            }
-
-            toolPropertiesMock.Setup(x => x.GetEvaluatedPropertyValue(It.IsAny<string>()))
-                .Returns<string>(s =>
-                {
-                    string propertyValue = null;
-                    projectItemConfig.fileConfigProperties?.TryGetValue(s, out propertyValue);
-                    return propertyValue ?? string.Empty;
-                });
-
-            var vcFileConfigMock = new Mock<VCFileConfiguration>();
-            vcFileConfigMock.SetupGet(x => x.Tool).Returns(toolPropertiesMock.Object);
-
-            return vcFileConfigMock.Object;
-        }
-
-        private static void AssertMessageLogged(Mock<ILogger> loggerMock, string message)
-        {
-            loggerMock.Verify(x => x.WriteLine(It.Is<string>(
-                s => s.Equals(message))), Times.Once);
-        }
-
-        private static void AssertPartialMessageLogged(Mock<ILogger> loggerMock, string message)
-        {
-            loggerMock.Verify(x => x.WriteLine(It.Is<string>(
-                s => s.Contains(message))), Times.Once);
         }
     }
 }
